reorganize headers into sub-folders
