+++ include/LDtkLoader/Enum.hpp
@@ -5,8 +5,8 @@
 #include <string>
 #include <vector>
 #include "LDtkLoader/thirdparty/json.hpp"
+#include "LDtkLoader/containers/TagsContainer.hpp"
 #include "LDtkLoader/DataTypes.hpp"
-#include "LDtkLoader/TagsContainer.hpp"
 
 namespace ldtk {
