+++ include/LDtkLoader/defs/EntityDef.hpp
@@ -5,8 +5,8 @@
 #include <string>
 
 #include "LDtkLoader/thirdparty/json.hpp"
+#include "LDtkLoader/containers/TagsContainer.hpp"
 #include "LDtkLoader/DataTypes.hpp"
-#include "LDtkLoader/TagsContainer.hpp"
 #include "LDtkLoader/Tileset.hpp"
 
 namespace ldtk {
